requirements.txt
Flask-Login==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
Flask>=2.0.2
celery
Redis
mysql-connector==2.2.9
Flask-Session
flask-mysqldb

